fix defaults for motion
